pydantic all the things (#149)
This turned into a GIANT PR, everything is working, but there seems to be an endless number of nice to haves. This needs to just be merged and let the nice to haves get their own PR.  It is currently working, but there may be some rough edges on the pre-release.